Move tests to `tests` folder as they are integration tests
